requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
pydantic==2.9.2
python-dotenv==1.0.1

# SQL & dbt
sqlglot==25.24.0
dbt-core==1.8.7
dbt-duckdb==1.8.4

# Lineage graph
networkx==3.3
matplotlib==3.9.2

# LLM
anthropic>=0.84.0

# Utilities
pandas==2.2.3
pytest==9.0.2
